+++ catalog/app/components/StackedAreaChart/StackedAreaChart.js
@@ -20,7 +20,6 @@ export default function MultiSparkline({
   axis = true,
   extendL = false,
   extendR = false,
-  fade = 0.5,
   padding = 1,
   px = padding,
   py = padding,
@@ -62,11 +61,15 @@ export default function MultiSparkline({
     [xScale, yScale, extendL, extendR, width],
   )
 
-  const areas = React.useMemo(
+  const figures = React.useMemo(
     () =>
       R.pipe(
         R.aperture(2),
-        R.map(([bottom, top]) => R.concat(mkPoints(top), R.reverse(mkPoints(bottom)))),
+        R.map(([bottom, top]) => ({
+          area: [...mkPoints(top), ...R.reverse(mkPoints(bottom))],
+          stroke: [...mkPoints(top), ...R.reverse(mkPoints(top))],
+        })),
+        R.reverse,
       )(stacked),
     [stacked, mkPoints],
   )
@@ -122,22 +125,26 @@ export default function MultiSparkline({
         {!!onCursor && cursorPaint.def}
       </defs>
       <g>
-        {areas.map(
-          (points, i) =>
-            points && (
-              <polygon
-                // eslint-disable-next-line react/no-array-index-key
-                key={`area:${i}`}
-                onMouseEnter={handleAreaEnter(i)}
-                onMouseLeave={handleAreaLeave}
-                points={SVG.pointsToSVG(points)}
-                fill={areaPaints[i].ref}
-                fillOpacity={
-                  onCursor && cursor && cursor.i != null && cursor.i !== i ? fade : 1
-                }
-              />
-            ),
-        )}
+        {figures.map(({ area, stroke }, i) => (
+          <g
+            // eslint-disable-next-line react/no-array-index-key
+            key={`area:${i}`}
+            onMouseEnter={handleAreaEnter(i)}
+            onMouseLeave={handleAreaLeave}
+            opacity={onCursor && cursor?.i != null && cursor.i !== i ? 0.4 : 1}
+          >
+            <polygon
+              points={SVG.pointsToSVG(area)}
+              fill={areaPaints[i].ref}
+              fillOpacity={onCursor && cursor?.i === i ? 0.6 : 0.5}
+            />
+            <polyline
+              points={SVG.pointsToSVG(stroke)}
+              stroke={areaPaints[i].ref}
+              strokeWidth={2}
+            />
+          </g>
+        ))}
         {!!cursor && !!onCursor && (
           <line
             x1={xScale(cursor.j)}