+++ catalog/app/containers/Bucket/errors.tsx
@@ -43,7 +43,7 @@ export class FileNotFound extends BucketError {}
 export class VersionNotFound extends BucketError {}
 
 export interface BucketPreferencesInvalidProps {
-  errors: { dataPath: string; message: string }[]
+  errors: { dataPath?: string; message?: string }[]
 }
 
 export class BucketPreferencesInvalid extends BucketError {
@@ -58,7 +58,7 @@ export class BucketPreferencesInvalid extends BucketError {
 }
 
 export interface WorkflowsConfigInvalidProps {
-  errors: { dataPath: string; message: string }[]
+  errors: { dataPath?: string; message?: string }[]
 }
 
 export class WorkflowsConfigInvalid extends BucketError {