+++ catalog/app/embed/Embed.js
@@ -41,13 +41,14 @@ import WithGlobalStyles from '../global-styles'
 
 import AppBar from './AppBar'
 import * as EmbedConfig from './EmbedConfig'
+import * as Overrides from './Overrides'
+import * as ipc from './ipc'
 
-const mkLazy = (load) =>
-  RT.loadable(load, { fallback: () => <Placeholder color="text.secondary" /> })
+const SuspensePlaceholder = () => <Placeholder color="text.secondary" />
 
-const Dir = mkLazy(() => import('./Dir'))
-const File = mkLazy(() => import('./File'))
-const Search = mkLazy(() => import('./Search'))
+const Dir = RT.mkLazy(() => import('./Dir'), SuspensePlaceholder)
+const File = RT.mkLazy(() => import('./File'), SuspensePlaceholder)
+const Search = RT.mkLazy(() => import('./Search'), SuspensePlaceholder)
 
 const FinalBoundary = createBoundary(() => (error) => (
   <h1
@@ -136,36 +137,40 @@ function BucketLayout({ bucket, children }) {
 }
 
 function useInit() {
+  const messageParent = ipc.useMessageParent()
   const [state, setState] = React.useState(null)
 
-  const handleMessage = React.useCallback(
-    ({ data }) => {
-      if (!data || data.type !== 'init') return
-      const { type, ...init } = data
-      try {
-        if (!init.bucket) throw new Error('missing .bucket')
-        if (init.scope) {
-          if (typeof init.scope !== 'string') throw new Error('.scope must be a string')
-          init.scope = s3paths.ensureSlash(init.scope)
+  ipc.useMessageHandler(
+    React.useCallback(
+      ({ type, ...init }) => {
+        if (type !== 'init') return
+        try {
+          if (!init.bucket && !init.route)
+            throw new Error('missing either .bucket or .route')
+          if (init.scope) {
+            if (typeof init.scope !== 'string') throw new Error('.scope must be a string')
+            // eslint-disable-next-line no-param-reassign
+            init.scope = s3paths.ensureSlash(init.scope)
+          }
+          Overrides.validate(init.overrides)
+          setState(init)
+        } catch (e) {
+          const message = `Configuration error: ${e.message}`
+          // eslint-disable-next-line no-console
+          console.error(message)
+          // eslint-disable-next-line no-console
+          console.log('init object:', init)
+          messageParent({ type: 'error', message, init })
+          setState(e)
         }
-        setState(init)
-      } catch (e) {
-        // eslint-disable-next-line no-console
-        console.error(`Configuration error: ${e.message}`)
-        // eslint-disable-next-line no-console
-        console.log('init object:', init)
-        setState(e)
-      }
-    },
-    [setState],
+      },
+      [setState, messageParent],
+    ),
   )
 
   React.useEffect(() => {
-    window.addEventListener('message', handleMessage)
-    return () => {
-      window.removeEventListener('message', handleMessage)
-    }
-  }, [handleMessage])
+    messageParent({ type: 'ready' })
+  }, [messageParent])
 
   return state
 }
@@ -191,6 +196,7 @@ function Init() {
 
 function usePostInit(init) {
   const dispatch = redux.useDispatch()
+  const messageParent = ipc.useMessageParent()
   const [state, setState] = React.useState(null)
 
   React.useEffect(() => {
@@ -199,6 +205,7 @@ function usePostInit(init) {
     result.promise
       .then(() => {
         setState(true)
+        messageParent({ type: 'init', init })
       })
       .catch((e) => {
         // eslint-disable-next-line no-console
@@ -206,8 +213,13 @@ function usePostInit(init) {
         // eslint-disable-next-line no-console
         console.error(e)
         setState(new ErrorDisplay('Authentication Failure'))
+        messageParent({
+          type: 'error',
+          message: `Authentication failure: ${e.message}`,
+          credentials: init.credentials,
+        })
       })
-  }, [init, dispatch])
+  }, [init, dispatch, messageParent])
 
   return state
 }
@@ -258,11 +270,52 @@ function useCssFiles(files = []) {
   }, [files])
 }
 
+function useSyncHistory(history) {
+  const messageParent = ipc.useMessageParent()
+
+  ipc.useMessageHandler(
+    React.useCallback(
+      ({ type, ...data }) => {
+        if (type !== 'navigate') return
+        try {
+          if (!data.route) throw new Error('missing .route')
+          if (typeof data.route !== 'string') throw new Error('.route must be a string')
+          history.push(data.route)
+        } catch (e) {
+          const message = `Navigate: error: ${e.message}`
+          // eslint-disable-next-line no-console
+          console.error(message)
+          // eslint-disable-next-line no-console
+          console.log('params:', data)
+          messageParent({ type: 'error', message, data })
+        }
+      },
+      [history, messageParent],
+    ),
+  )
+
+  React.useEffect(
+    () =>
+      history.listen((l, action) => {
+        messageParent({
+          type: 'navigate',
+          route: `${l.pathname}${l.search}${l.hash}`,
+          action,
+        })
+      }),
+    [history, messageParent],
+  )
+}
+
 function App({ init }) {
   const { urls } = NamedRoutes.use()
   const history = useConstant(() =>
-    createHistory({ initialEntries: [urls.bucketDir(init.bucket, init.path)] }),
+    createHistory({
+      initialEntries: [init.route || urls.bucketDir(init.bucket, init.path)],
+    }),
   )
+
+  useSyncHistory(history)
 
   const storage = useConstant(() => ({
     load: () => ({}),
@@ -273,6 +326,7 @@ function App({ init }) {
   useCssFiles(init.css)
 
   return RT.nest(
+    [Overrides.Provider, { value: init.overrides }],
     [EmbedConfig.Provider, { config: init }],
     [CustomThemeProvider, { theme: init.theme }],
     [Store.Provider, { history }],