+++ catalog/app/embed/File.js
@@ -32,7 +32,17 @@ import renderPreview from 'containers/Bucket/renderPreview'
 import * as requests from 'containers/Bucket/requests'
 
 import * as EmbedConfig from './EmbedConfig'
+import * as Overrides from './Overrides'
 import getCrumbs from './getCrumbs'
+import * as ipc from './ipc'
+
+const defaults = {
+  s3ObjectLink: {
+    title: "Copy object version's canonical HTTPS URI to the clipboard",
+    href: (ctx) => ctx.s3HttpsUri,
+    notification: 'HTTPS URI copied to clipboard',
+  },
+}
 
 const useVersionInfoStyles = M.makeStyles(({ typography }) => ({
   version: {
@@ -54,6 +64,7 @@ function VersionInfo({ bucket, path, version }) {
   const { urls } = NamedRoutes.use()
   const cfg = Config.use()
   const { push } = Notifications.use()
+  const messageParent = ipc.useMessageParent()
 
   const containerRef = React.useRef()
   const [anchor, setAnchor] = React.useState()
@@ -61,16 +72,37 @@ function VersionInfo({ bucket, path, version }) {
   const open = React.useCallback(() => setOpened(true), [])
   const close = React.useCallback(() => setOpened(false), [])
 
+  const overrides = Overrides.use(defaults)
+
   const classes = useVersionInfoStyles()
 
-  const getHttpsUri = (v) =>
-    s3paths.handleToHttpsUri({ bucket, key: path, version: v.id })
+  const getLink = (v) =>
+    overrides.s3ObjectLink.href({
+      url: urls.bucketFile(bucket, path, v.id),
+      s3HttpsUri: s3paths.handleToHttpsUri({ bucket, key: path, version: v.id }),
+      bucket,
+      key: path,
+      version: v.id,
+    })
+
   const getCliArgs = (v) => `--bucket ${bucket} --key "${path}" --version-id ${v.id}`
 
-  const copyHttpsUri = (v) => (e) => {
+  const copyLink = (v) => (e) => {
     e.preventDefault()
-    copyToClipboard(getHttpsUri(v), { container: containerRef.current })
-    push('HTTPS URI copied to clipboard')
+    if (overrides.s3ObjectLink.emit !== 'override') {
+      copyToClipboard(getLink(v), { container: containerRef.current })
+      push(overrides.s3ObjectLink.notification)
+    }
+    if (overrides.s3ObjectLink.emit) {
+      messageParent({
+        type: 's3ObjectLink',
+        url: urls.bucketFile(bucket, path, v.id),
+        s3HttpsUri: s3paths.handleToHttpsUri({ bucket, key: path, version: v.id }),
+        bucket,
+        key: path,
+        version: v.id,
+      })
+    }
   }
 
   const copyCliArgs = (v) => (e) => {
@@ -145,9 +177,9 @@ function VersionInfo({ bucket, path, version }) {
                         )}
                       <M.Hidden xsDown>
                         <M.IconButton
-                          title="Copy object version's canonical HTTPS URI to the clipboard"
-                          href={getHttpsUri(v)}
-                          onClick={copyHttpsUri(v)}
+                          title={overrides.s3ObjectLink.title}
+                          href={getLink(v)}
+                          onClick={copyLink(v)}
                         >
                           <M.Icon>link</M.Icon>
                         </M.IconButton>
@@ -444,7 +476,7 @@ export default function File({
                   Err: (e) => {
                     throw e
                   },
-                  Ok: withPreview(renderPreview),
+                  Ok: withPreview(renderPreview()),
                 })}
               </Section>
               <Meta bucket={bucket} path={path} version={version} />